Move Shape to abstractProvisioner.py to avoid cyclic import involving common.py
This will make more sense once we add the getNodeShape method to AbstractProvisioner
